Split the app.py int multiple files
